refactor(auth): split 754-line OAuth handler into 3 files
oauth_providers.go: List/Initiate handlers + deps + route registration
oauth_callback.go: Callback + token exchange + helpers
oauth_linking.go: Link init/accounts/unlink/status handlers
